Move getRepeatScheduleOptions
